+++ src/libc/errno.rs
@@ -12,17 +12,18 @@ use crate::Environment;
 
 pub const EPERM: i32 = 1;
 pub const EDEADLK: i32 = 11;
+pub const EBUSY: i32 = 16;
 pub const EINVAL: i32 = 22;
 
 #[derive(Default)]
 pub struct State {
-    errnos: std::collections::HashMap<crate::ThreadID, MutPtr<i32>>,
+    errnos: std::collections::HashMap<crate::ThreadId, MutPtr<i32>>,
 }
 impl State {
     fn errno_for_thread(
         &mut self,
         mem: &mut crate::mem::Mem,
-        thread: crate::ThreadID,
+        thread: crate::ThreadId,
     ) -> MutPtr<i32> {
         *self.errnos.entry(thread).or_insert_with(|| {
             log!(